fix(share): type Share.expire as Double for @JsExport compatibility
Kotlin/JS refuses to @JsExport classes containing Long fields unless
the compilation opts into -Xes-long-as-bigint or an es2015 target, and
the first CI run on this branch flagged Share.expire accordingly. The
rest of krfiles already sidesteps this by representing numeric fields
as Double (see Resource.size / Resource.mode); the change here just
aligns Share with the same convention. Filebrowser share-link expiries
live well within the 2^53 safe-integer range, so the round-trip is
lossless for any realistic value.

The vendored cli/native/libkrfiles_api.h is regenerated so the shim
still compiles cleanly — the only effective change there is that
Share.get_expire / Share(...) now take a KDouble instead of a KLong.

+++ lib/src/commonMain/kotlin/dev/rolandh/krfiles/Models.kt
@@ -161,6 +161,8 @@ internal data class ActionRequest(
  * @property path Server-side path the share points at (what was shared).
  * @property userID ID of the user who created the share.
  * @property expire Expiry as a Unix epoch in seconds. `0` means no expiry.
+ *   Typed as [Double] (not [Long]) to stay compatible with the JS target's
+ *   `@JsExport` rules — for timestamps up to year 287396 this is lossless.
  * @property passwordHash bcrypt hash of the share password, empty if unset.
  * @property token Secondary token required to download password-protected
  *   shares via query arg. Empty when no password is set.
@@ -171,7 +173,7 @@ public data class Share(
     val hash: String = "",
     val path: String = "",
     val userID: Int = 0,
-    val expire: Long = 0,
+    val expire: Double = 0.0,
     @kotlinx.serialization.SerialName("password_hash")
     val passwordHash: String = "",
     val token: String = "",
